Spliting the `register` module into multiple files
