+++ pulser/devices/_device_datacls.py
@@ -21,7 +21,7 @@ import numpy as np
 from scipy.spatial.distance import pdist, squareform
 
 from pulser import Pulse
-from pulser.register import BaseRegister
+from pulser.register.base_register import BaseRegister
 from pulser.channels import Channel
 from pulser.json.utils import obj_to_dict
 from pulser.devices.interaction_coefficients import c6_dict
